feat: enhance block compression and shader performance
- Added unique ID tracking in block compression to optimize palette generation.
- Refactored OpaqueShader and TransparentShader to use axis face encoding for normals, reducing interpolation costs.
- Improved AO pipeline by simplifying position calculations for adjacent cells.
- Changed frustum plane cache to Float32Array for better performance.
- Updated MeshSerializer documentation to clarify gzip compression usage.
- Introduced live data size tracking in OpfsChunkStore for efficient memory management.
- Enhanced OPFS worker with a ring buffer for serial operation queue, improving async handling.
- Implemented LRU caching for region files to optimize memory usage.
- Added PineLeaves block type to BlockType enum for new texture support.

+++ src/code/World/Occlusion/OcclusionCuller.ts
@@ -68,9 +68,9 @@ const _incBfsEntry = new Int8Array(BFS_CAP);
 const _incBfsSteps = new Uint16Array(BFS_CAP);
 
 // ---------------------------------------------------------------------------
-// Frustum plane cache — packed Float64Array
+// Frustum plane cache — packed Float32Array
 // ---------------------------------------------------------------------------
-const _frustumPacked = new Float64Array(24);
+const _frustumPacked = new Float32Array(24);
 let _frustumValid = false;
 let _lastVPHash = -1;
